% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preknit_posthtmlrender.R
\name{eval_replace}
\alias{eval_replace}
\title{Evaluate and replace code chunks/lines}
\usage{
eval_replace(infile, symbol, outfile = NULL, fileext = NULL,
  encoding = "UTF-8")
}
\arguments{
\item{infile}{- input Rmd or html file}

\item{symbol}{- symbol string to identify which are code chunks/lines to processed}

\item{outfile}{- output file name after processing. Either outfile or fileext must be provided.}

\item{fileext}{- file extension of output file is required if outfile is not provided.}

\item{encoding}{- file encoding is required to readLines of infile}
}
\description{
Extract code chunks for Rmd or code lines for html, evaluate each line and
replace original code chunks/lines with captured output in original positions
}

