% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preknit_posthtmlrender.R
\name{eval_src}
\alias{eval_src}
\title{Eval a character array of code}
\usage{
eval_src(srcCharVec, ...)
}
\arguments{
\item{srcCharVec}{- code character array}

\item{...}{- arguments to be passed on to other functions}
}
\description{
Evaluate an array of code line by line and capture output
}

