% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preknit_posthtmlrender.R
\name{post_html_render_proc}
\alias{post_html_render_proc}
\title{Post HTML-Rendering Processing}
\usage{
post_html_render_proc(inhtml, outhtml = NULL)
}
\arguments{
\item{inhtml}{- input file for be processed}

\item{outhtml}{- output file name}
}
\description{
Evaluates code chunks in between `%%% MY_CODE_HERE `. Replace
these code chunks with evaluated output captured using capture.output function.
Typically, we want to insert an external HTML file and we can
use `%%% writeLines(readLines("MY_EXTERNAL_HTML_FILE"))` within the Rmd file.
This function will read in the external html file and replace this `%%% ` with
contents in the html file.
}

