% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preknit_posthtmlrender.R
\name{run_code}
\alias{run_code}
\title{Eval a line of code and capture output}
\usage{
run_code(oneline, envir)
}
\arguments{
\item{oneline}{- code line}

\item{envir}{- environment to evaluate the code in}
}
\description{
Run a line of code and capture output into a character string
}

