% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preknit_posthtmlrender.R
\name{pre_knit_proc}
\alias{pre_knit_proc}
\title{Pre-knitting Processing}
\usage{
pre_knit_proc(infile, outfile = NULL)
}
\arguments{
\item{infile}{- input file for be processed}

\item{outfile}{- output file name}
}
\description{
Evaluates code chunks which are marked as between adjacent lines of @@s.
Replace these code chunks with evaluated output captured using capture.output function.
}

